refactor: Phase 3 architectural refinement — decompose server monolith, merge bulk manager, fix version source of truth
## Server Decomposition (778 → ~210 lines)
- Extract server.py monolith into server/ package with modular routers:
  - server/routes/config.py — /api/config endpoints
  - server/routes/clusters.py — /api/clusters, /api/cluster/{name} endpoints
  - server/routes/indices.py — unhealthy index diagnostics and fix actions
  - server/routes/cli.py — subprocess CLI proxy
  - server/schemas.py — Pydantic request/response models (single source)
  - server/services.py — shared build_es_client() factory and parse_index_size()
- Eliminates 4x copy-pasted client construction blocks
- Eliminates inline size-parsing logic duplication
- Tightens CORS from wildcard (*) to http://127.0.0.1 (OWASP hardened)

## BulkDocumentManager Consolidation
- Add bulk_index_sync() and bulk_delete_sync() to DocumentManager (BaseManager)
- Fix doc.pop('_id') mutation bug — now uses doc.copy() to preserve caller data
- New methods use _ensure_connected() and _handle_response() guards (BaseManager)
- Deprecate BulkDocumentManager as thin shim with DeprecationWarning
- Rewrite tests: 18 tests covering both new methods and deprecated shim

## Version Source of Truth
- Replace hardcoded __version__ in __init__.py with importlib.metadata.version()
- pyproject.toml is now the single canonical version source
- CI sed command still updates pyproject.toml; runtime reads from metadata

## Verification
- mypy: 0 errors across 66 source files
- black: 66 files compliant
- pytest: 54 failures (pre-existing baseline), 263 passed, 0 new regressions
- Live: CLI + GUI verified against local-elastro-brain docker cluster
