+++ elastro/__init__.py
@@ -4,7 +4,13 @@ Elasticsearch Management Module.
 A module for managing Elasticsearch operations within a pipeline process.
 """
 
-__version__ = "1.3.53"
+from importlib.metadata import version, PackageNotFoundError
+
+try:
+    __version__ = version("elastro-client")
+except PackageNotFoundError:
+    # Fallback for editable installs or when metadata is unavailable
+    __version__ = "0.0.0-dev"
 
 # Core component imports
 from elastro.core.client import ElasticsearchClient
